requirements.txt
python-telegram-bot==21.10
httpx==0.28.1
SQLAlchemy==2.0.36
python-docx==1.1.2
PyPDF2==3.0.1
weasyprint==63.1
Jinja2==3.1.5
